Move HttpClientTrait to http package
